+++ Hast.Transformer/Configuration/MemberInvocationInstanceCountConfiguration.cs
@@ -54,7 +54,7 @@ namespace Hast.Common.Configuration
         }
 
         // But why do I need to cast to uint? http://stackoverflow.com/questions/10065287/why-is-ushort-ushort-equal-to-int#comment58098182_10157517
-        public uint MaxInvocationInstanceCount { get { return (uint)((MaxRecursionDepth + 1) * MaxDegreeOfParallelism); } }
+        public int MaxInvocationInstanceCount { get { return (MaxRecursionDepth + 1) * MaxDegreeOfParallelism; } }
 
 
         /// <summary>
